+++ include/GafferImage/ImageStats.h
@@ -81,13 +81,21 @@ class GAFFERIMAGE_API ImageStats : public Gaffer::ComputeNode
 
 	protected :
 
-		/// Implemented to hash the area we are sampling along with the channel context and regionOfInterest.
 		void hash( const Gaffer::ValuePlug *output, const Gaffer::Context *context, IECore::MurmurHash &h ) const override;
-
-		/// Computes the min, max and average plugs by analyzing the input ImagePlug.
 		void compute( Gaffer::ValuePlug *output, const Gaffer::Context *context ) const override;
+		Gaffer::ValuePlug::CachePolicy computeCachePolicy( const Gaffer::ValuePlug *output ) const override;
+		Gaffer::ValuePlug::CachePolicy hashCachePolicy( const Gaffer::ValuePlug *output ) const override;
+
 
 	private :
+
+		// Stats for individual tiles
+		Gaffer::ObjectPlug *tileStatsPlug();
+		const Gaffer::ObjectPlug *tileStatsPlug() const;
+
+		// Combined stats, before they get broken out into 3 seperate plugs
+		Gaffer::ObjectPlug *allStatsPlug();
+		const Gaffer::ObjectPlug *allStatsPlug() const;
 
 		// Input plug to receive the flattened image from the internal
 		// DeepState plug.
@@ -97,8 +105,6 @@ class GAFFERIMAGE_API ImageStats : public Gaffer::ComputeNode
 		// The internal DeepState node.
 		GafferImage::DeepState *deepState();
 		const GafferImage::DeepState *deepState() const;
-
-		std::string channelName( int colorIndex ) const;
 
 		static size_t g_firstPlugIndex;